+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `ArrayBytesFixedDisjointView[CreateError]`
 - Add support for data type extensions with `zarrs_data_type` 0.2.0
 - Add `custom_data_type_fixed_size` and `custom_data_type_variable_size` examples
+- Add `[Async]ArrayDlPackExt` traits that add methods to `Array` for `DLPack` tensor interop
+  - Gated by the `dlpack` feature
 
 ### Changed
 - **Breaking**: change `ArraySubset::inbounds` to take another subset rather than a shape
@@ -40,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - **Breaking**: `StorageTransformerPlugin` now uses a `Plugin`
 - Add `DataTypeExtension` variant to `CodecError`
 - `ArrayCreateError::DataTypeCreateError` now uses a `PluginCreateError` internally
+- **Breaking**: `ArrayError` is now marked as non-exhaustive
 
 ### Fixed
 - Fixed reserving one more element than necessary when retrieving `string` or `bytes` array elements